Add discord's media domain to blocklist-discord.txt
Discord has migrated to using around 9999 per server location of subdomains of the .discord.media domain to facilitate connecting to voice calls and voice channels, even when the main application is not accessible.

Since 2023 Pihole supports limited ABP style blocklists. This commit makes use of this feature to block all discord.com domains and all discord.media subdomains, ensuring the application is completely and properly blocked, even when the app was opened once on another network.